Refactoring the App Localization tool
In this PR, I'm refactoring the way nylo manage the translations.  
Because of dart doesn't populate the stringified json with quotes on the keys and values, we have to store the whole json in the _localizedStrings variable and iterate over the keys to get the plain values. This way we can support deeper keys instead of the actual deep level which one support only one key.

Open to discuss about it.